Make `#![no_std]` support an opt-in feature.
Abstracting away the difference between `std` and `core` for all the
places that use it is... difficult.  Importing one crate under the
other's name, as here, doesn't work unless it's reexported so the
macro expansion can use it, which isn't ideal.  Trying to use a macro
to expand to the right crate name, or prepend it to a given path,
breaks for reasons that I don't know the macro system well enough to
understand.

Also, this will become obsolete once Rust 1.6 ships with stable
`no_std`/`::core`.